Revert "Màj rechercheLocation"
This reverts commit cdb2131cf3a5bd37f2fcbdde26ea7cf401b06f50.

+++ src/main/java/fr/formation/afpa/dao/ILocationDao.java
@@ -10,5 +10,5 @@ import fr.formation.afpa.domain.Location;
 @Repository
 public interface ILocationDao extends JpaRepository<Location, Integer> {
 	
-	public List<Location> findBymaxColocataireLessThanEqualAndLoyerLessThanEqualAndSuperficeLessThanEqual(Integer maxColocataire, Integer loyer, Integer superfice);
+	public List<Location> findBymaxColocataireLessThanEqualAndLoyerLessThanEqualAndSuperficeLessThanEqual(Integer maxColocataire, Double loyer, Integer superfice);
 }
